Extract and fix transfer tests
