+++ src/lib.rs
@@ -18,6 +18,7 @@ mod unittest_private_metadata;
 mod unittest_public_metadata;
 mod unittest_queries;
 mod unittest_reveal;
+mod unittest_transfer;
 mod unittest_whitelisted_approval;
 mod utils;
 mod viewing_key;
